Fix multiple RTCViews, mirror & objectFit
VideoRendererGui is a singleton and, consequently, multiple RTCViews
would not render. Implements RTCView using SurfaceViewRenderer instead
in order to render multiple RTCView at a time.

Introduces zOrder in RTCView in order to allow control over how multiple
RTCViews overlap.

Introduces mirror in RTCView in order to allow control over whether
RTCView mirrors the video it renders. Commonly used for the user-facing
camera.

Implements objectFit in RTCView in order to allow control how the video
is displayed in the rectangle of RTCView while preserving the aspect
ratio. The feature was previously implemented on iOS.
on